+++ src/main/java/ru/pgk/smartbudget/features/expenseCategory/dto/ExpenseCategoryDto.java
@@ -1,6 +1,6 @@
 package ru.pgk.smartbudget.features.expenseCategory.dto;
 
 public record ExpenseCategoryDto(
-   Short id,
+   Integer id,
    String name
 ) {}
